Platform Obstacles
Made a few nested prefabs, each a set of platforms with the middle platform moving in a certain way. There are three types of middle platform movements so far:

-Moving back and fourth between platforms on X axis
-Moving back and fourth between platforms on Z axis
-Rotating

The data for each nested prefabs moving platform is received from a scriptable object, so if you want to mess with the speed, create another PlatformData SO instance, set the values and drag it into the desired moving platforms script, in the inspector.

You can drag a nested prefab into the scene and it will do its thing. Feel free to mess about with instances of a prefab ect ect.