Audio Fmod framework
Groundwork added for fmod events, only a few basic ones on the project at the moment, but a system is in place to be able to easily add events.

Current audio Events:

-PlayerJump (trigged in player controller script)
-Game music (triggered when scene loads) Note: transition sections have been made in fmod, so you will only hear the intro of music looped, once collectable origami is added music will progress with the amount of origami collected.

There are a few more sounds ready to go in fmod, for when we have more stuff implemented e.g. footstep sounds linked to animation frames.